Split MultiplicationView into separate files and added SwiftDoc comments
